+++ Runtime/Scripts/Managers/CursorManager.cs
@@ -6,13 +6,14 @@
  * Date Created: August 22, 2022
  * 
  * Additional Comments:
- *		File Line Length: 120
+ *      File Line Length: 120
  */
 
 using System.Collections;
 using System.Collections.Generic;
 using UnityEngine;
 
+using UnityEngine.InputSystem;
 using Kokowolo.Utilities;
 
 [DefaultExecutionOrder(-100)]
@@ -44,14 +45,20 @@ public class CursorManager : MonoSingleton<CursorManager>
 
     public static bool HasValidHitInfo => HitInfo.transform != null;
 
+    static bool doRaycast = true;
+
     #endregion
     /************************************************************/
     #region Functions
 
     private void LateUpdate() 
     {
-        Raycasting.RaycastFromMouseScreenPoint(out hitInfo, layerMask);
-        cursorVisual.transform.position = hitInfo.point;
+        if (doRaycast)
+        {
+            Raycasting.RaycastFromMouseScreenPoint(out hitInfo, layerMask);
+            cursorVisual.transform.position = hitInfo.point;
+        }
+        doRaycast = true;
     }
 
     public static void SetActiveCursorVisual(bool value)
@@ -62,6 +69,16 @@ public class CursorManager : MonoSingleton<CursorManager>
     public static Vector3 GetWorldPosition()
     {
         return Instance.hitInfo.point;
+    }
+
+    public static void SetCursorPosition(Vector3 worldPosition)
+    {
+        Vector3 screenPoint = Camera.main.WorldToScreenPoint(worldPosition);
+        Mouse.current.WarpCursorPosition(screenPoint);
+
+        Instance.hitInfo.point = worldPosition;
+        Instance.cursorVisual.transform.position = worldPosition;
+        doRaycast = false;
     }
 
     #endregion